tools/images_gen/requirements.txt
pycairo>=1.20.0
Pillow>=9.0.0
